producer/produce_data/requirements.txt
requests==2.28.1
kafka-python==2.0.2
python-schema-registry-client==2.4.1
faust-streaming==0.10.4